moved all site sections to individual components, removed unused files, updated css for contact form
