+++ src/App.css
@@ -4,176 +4,9 @@
   /* border: 1px red solid; */
 }
 
-body {
-  /* background-color: rgba(0, 0, 255, 0.2); */
-}
-
 .App {
   display: flex;
   flex-direction: column;
   justify-content: flex-end;
   align-items: center;
-}
-
-.jump-link {
-  height: 0px;
-  width: 100%;
-  visibility: hidden;
-  z-index: -1;
-  padding-top: 86px;
-  margin-top: -86px;
-  background-clip: content-box;
-}
-
-.company {
-  display: flex;
-  flex-direction: row;
-}
-
-.company-name {
-  padding-left: 1rem;
-}
-
-.top-navigation {
-  width: 25%;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-evenly;
-  font-weight: bold;
-}
-
-.lightgray-background {
-  background-color: #0000000d;
-}
-
-.section-a {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-}
-
-.section-a > .section-b {
-  width: 50%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.section-c {
-  text-align: center;
-}
-
-.cards-container {
-  display: flex;
-  flex-direction: row;
-  width: 80%;
-  justify-content: space-between;
-  flex-wrap: wrap;
-}
-
-.card {
-  display: flex;
-  flex-direction: column;
-  width: 23%;
-  align-items: center;
-  padding-bottom: 2rem;
-}
-
-.card-title {
-  max-width: 15rem;
-  width: 15rem;
-  height: 15rem;
-  border-radius: 1rem;
-}
-
-.card-title-image {
-  position: absolute;
-  width: inherit;
-  height: inherit;
-  object-fit: cover;
-  border-radius: inherit;
-  filter: brightness(0.5);
-}
-
-.card-title-text {
-  position: relative;
-  color: white;
-  width: inherit;
-  height: inherit;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.card-description {
-  text-align: center;
-  width: 90%;
-}
-
-.cityscape-image {
-  background-image: url("../public/cityscape.jpg");
-  width: 100%;
-  height: 25rem;
-  background-repeat: no-repeat;
-  background-size: cover;
-  background-position-y: center;
-  filter: brightness(1.3) opacity(0.33);
-  background-attachment: fixed;
-}
-
-.contact {
-  padding-bottom: 3rem;
-}
-
-.contact-form {
-  display: flex;
-  flex-direction: column;
-  width: 40%;
-  align-items: flex-start;
-  justify-content: space-between;
-}
-
-.input-package {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-}
-
-.input {
-  width: 100%;
-  border: 1px lightgray solid;
-  border-radius: 1.5rem;
-  padding: 1rem;
-}
-
-.message {
-  resize: none;
-  overflow: hidden;
-  padding: 1rem;
-  border-radius: 1.5rem;
-  border: 1px lightgray solid;
-  width: 100%;
-  height: auto;
-  margin-bottom: 1rem;
-}
-
-.error-message:has(+ [aria-invalid="true"]) {
-  visibility: visible;
-  color: red;
-  height: 1.2rem;
-}
-
-[aria-invalid="true"] {
-  border: 1px solid red;
-}
-
-[aria-disabled="true"].contact-me-button {
-  cursor: not-allowed;
-  filter: opacity(0.5);
-}
-
-.error-message {
-  visibility: hidden;
-  height: 0.6rem;
 }
