--- src/Playground.css
@@ -1,12 +1,0 @@
-.playground {
-    height: 100vh;
-    width: 100vw;
-    background-color: rgba(0, 0, 255, 0.2);
-    overflow: hidden;
-    -webkit-touch-callout: none;
-    -webkit-user-select: none;
-    -khtml-user-select: none;
-    -moz-user-select: none;
-    -ms-user-select: none;
-    user-select: none;
-}
