--- src/WebsiteSection.css
@@ -1,70 +1,0 @@
-.website-section {
-  width: 60%;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  /* background-image: linear-gradient(90deg, rgba(255,255,204,.2), transparent); */
-}
-
-.description {
-  width: 25rem;
-  /* filter: drop-shadow(0em 0em 1px white); */
-}
-
-.section {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  /* background-image: linear-gradient(270deg, rgba(255,255,204,.2), transparent); */
-  padding-bottom: 2rem;
-  background-color: #0000ff33;
-}
-
-.section-header {
-  width: 50%;
-  text-align: center;
-  font-size: x-large;
-  line-height: 3.5rem;
-}
-
-.form-submit {
-  width: 100%;
-  display: flex;
-  align-items: baseline;
-  flex-direction: row;
-  justify-content: flex-start;
-  column-gap: 1.5rem;
-}
-
-.contact-me-button {
-  padding-left: 1rem;
-  padding-right: 1rem;
-  padding-top: 0.6rem;
-  padding-bottom: 0.6rem;
-  text-decoration: none;
-  border-radius: 2rem;
-  background-color: rgb(206, 37, 85);
-  color: white;
-  font-weight: bold;
-  font-size: medium;
-  border: none;
-  filter: drop-shadow(2px 2px 2px black);
-}
-
-.form-submit-message {
-    font-weight: bold;
-}
-
-.hidden {
-    visibility: hidden;
-    opacity: 0;
-    transition: visibility 2s ease-out, opacity 2s linear;
-}
-
-.visible {
-    visibility: visible;
-    opacity: 1;
-    transition: visibility 1s ease-out, opacity 1s linear;
-}
