+++ src/about/About.css
@@ -10,4 +10,5 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
+  padding-bottom: 1rem;
 }
