+++ src/global.css
@@ -1,4 +1,8 @@
-.global-contact-me-button {
+h1 + p {
+    margin-block-start: unset;
+}
+
+.global-button {
   padding-left: 1rem;
   padding-right: 1rem;
   padding-top: 0.6rem;
@@ -29,4 +33,9 @@
 
 .global-text-align-center {
   text-align: center;
+}
+
+.global-text-link {
+  font-weight: bold;
+  color: rgba(0, 0, 255, 1);
 }
