+++ src/services/Services.js
@@ -1,0 +1,92 @@
+import "./Services.css";
+
+export default function Services() {
+  return (
+    <div className="services">
+      <div id="services-landing" className="global-jump-link"></div>
+      <h1>Services</h1>
+      <div className="services-cards-container">
+        <div className="services-card">
+          <div className="services-card-title">
+            <img className="services-card-title-image" src="gummy_worms.jpg" />
+            <div className="services-card-title-text">Custom Software</div>
+          </div>
+          <div className="services-card-description">
+            Tailored software solutions that perfectly fit your unique goals.
+          </div>
+        </div>
+        <div className="services-card">
+          <div className="services-card-title">
+            <img
+              className="services-card-title-image"
+              src="containered_assorted_candy.jpg"
+            />
+            <div className="services-card-title-text">Website Design</div>
+          </div>
+          <div className="services-card-description">
+            Visually stunning and user-friendly websites that represent your
+            brand effectively.
+          </div>
+        </div>
+        <div className="services-card">
+          <div className="services-card-title">
+            <img className="services-card-title-image" src="gummy_bears.jpg" />
+            <div className="services-card-title-text">Mobile Apps</div>
+          </div>
+          <div className="services-card-description">
+            Portable device solutions that connect you with your customers
+            anytime and anywhere.
+          </div>
+        </div>
+        <div className="services-card">
+          <div className="services-card-title">
+            <img className="services-card-title-image" src="m_ms.jpg" />
+            <div className="services-card-title-text">E-Commerce Solutions</div>
+          </div>
+          <div className="services-card-description">
+            Establish an e-commerce presense with trusted platforms with your
+            bottom line in mind.
+          </div>
+        </div>
+        <div className="services-card">
+          <div className="services-card-title">
+            <img className="services-card-title-image" src="assorted_candy.jpg" />
+            <div className="services-card-title-text">System Integration</div>
+          </div>
+          <div className="services-card-description">
+            Convert disjointed systems into a cohesive and effective
+            singularity.
+          </div>
+        </div>
+        <div className="services-card">
+          <div className="services-card-title">
+            <img className="services-card-title-image" src="jelly_beans.jpg" />
+            <div className="services-card-title-text">SEO Services</div>
+          </div>
+          <div className="services-card-description">
+            Enhance online visibility and drive traffic from leading search
+            engines.
+          </div>
+        </div>
+        <div className="services-card">
+          <div className="services-card-title">
+            <img className="services-card-title-image" src="spirals.jpg" />
+            <div className="services-card-title-text">Support & Maintenance</div>
+          </div>
+          <div className="services-card-description">
+            Keep applications running smoothly at all times.
+          </div>
+        </div>
+        <div className="services-card">
+          <div className="services-card-title">
+            <img className="services-card-title-image" src="macaroons.jpg" />
+            <div className="services-card-title-text">Consultation</div>
+          </div>
+          <div className="services-card-description">
+            Insightful guides for your technology initiatives.
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
